@@ -1,7 +1,7 @@
 {
     "name": "ngx-material-timepicker",
     "description": "Handy material design timepicker for angular",
-    "version": "2.11.2",
+    "version": "4.0.2",
     "license": "MIT",
     "author": "Vitalii Boiko <boyko330@gmail.com>",
     "keywords": [
@@ -11,6 +11,7 @@
         "angular 5",
         "angular 6",
         "angular 7",
+        "angular 8",
         "material design",
         "timepicker",
         "material timepicker",
@@ -31,7 +32,7 @@
         "build": "ng build --prod",
         "lint": "ng lint",
         "test": "ng test --watch=false --code-coverage --browsers ChromeHeadlessNoSandbox",
-        "test:watch": "ng test --code-coverage",
+        "test:watch": "ng test --watch=true --browsers ChromeHeadlessNoSandbox --code-coverage",
         "codecov": "codecov",
         "packagr": "ng-packagr -p ng-package.json",
         "publish": "run-s packagr && npm publish dist",
@@ -39,28 +40,28 @@
     },
     "types": ".src/app/material-timepicker/index.d.ts",
     "private": false,
-    "peerDependencies": {
-        "moment": "^2.21.0"
+    "dependencies": {
+        "luxon": "^1.11.3"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.6.5",
-        "@angular/animations": "^7.0.3",
-        "@angular/cli": "^7.0.5",
-        "@angular/common": "^7.0.3",
-        "@angular/compiler": "^7.0.3",
-        "@angular/compiler-cli": "^7.0.3",
-        "@angular/core": "^7.0.3",
-        "@angular/forms": "^7.0.3",
-        "@angular/http": "^7.0.3",
-        "@angular/language-service": "^7.0.3",
-        "@angular/platform-browser": "^7.0.3",
-        "@angular/platform-browser-dynamic": "^7.0.3",
-        "@angular/router": "^7.0.3",
+        "@angular-devkit/build-angular": "~0.800.0",
+        "@angular/animations": "^8.0.0",
+        "@angular/cli": "^8.0.0",
+        "@angular/common": "^8.0.0",
+        "@angular/compiler": "^8.0.0",
+        "@angular/compiler-cli": "^8.0.0",
+        "@angular/core": "^8.0.0",
+        "@angular/forms": "^8.0.0",
+        "@angular/language-service": "^8.0.0",
+        "@angular/platform-browser": "^8.0.0",
+        "@angular/platform-browser-dynamic": "^8.0.0",
+        "@angular/router": "^8.0.0",
         "@types/jasmine": "~2.8.3",
         "@types/jasminewd2": "~2.0.2",
+        "@types/luxon": "^1.11.1",
         "@types/node": "~6.0.60",
         "codecov": "^3.1.0",
-        "codelyzer": "^4.0.1",
+        "codelyzer": "^5.0.1",
         "core-js": "^2.4.1",
         "jasmine-core": "~2.8.0",
         "jasmine-spec-reporter": "~4.2.1",
@@ -69,17 +70,18 @@
         "karma-coverage-istanbul-reporter": "^1.2.1",
         "karma-jasmine": "~1.1.0",
         "karma-jasmine-html-reporter": "^0.2.2",
-        "ng-packagr": "^4.4.0",
-        "ngx-material-timepicker": "^2.8.3",
+        "luxon": "^1.11.3",
+        "ng-packagr": "^5.2.0",
+        "ngx-material-timepicker": "^3.2.2",
+        "node-sass": "^4.12.0",
         "npm-run-all": "^4.1.3",
-        "protractor": "~5.1.2",
-        "rxjs": "^6.3.3",
+        "protractor": "^5.4.2",
+        "rxjs": "^6.5.2",
         "travis-deploy-once": "^5.0.0",
         "ts-node": "~4.1.0",
-        "tsickle": "^0.34.0",
+        "tsickle": "^0.35.0",
         "tslint": "~5.9.1",
-        "typescript": "3.1.6",
-        "zone.js": "^0.8.26",
-        "moment": "^2.21.0"
+        "typescript": "3.4.5",
+        "zone.js": "~0.9.1"
     }
 }
