@@ -1,5 +1,6 @@
-import {Component} from '@angular/core';
-import {NgxMaterialTimepickerTheme} from 'ngx-material-timepicker';
+/* tslint:disable:max-line-length */
+import { Component } from '@angular/core';
+import { NgxMaterialTimepickerTheme } from './material-timepicker/models/ngx-material-timepicker-theme.interface';
 
 @Component({
     selector: 'app-root',
@@ -7,7 +8,6 @@ import {NgxMaterialTimepickerTheme} from 'ngx-material-timepicker';
     styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-
 
     simpleExamples = `
        <div class="12hr-example">
@@ -18,18 +18,18 @@ export class AppComponent {
        <div class="24hr-example">
             <input placeholder="24hr format" aria-label="24hr format" [ngxTimepicker]="fullTime" [format]="24"  readonly>
             <ngx-material-timepicker #fullTime></ngx-material-timepicker>
-       </div>       
+       </div>
        <div class="default-time-example">
             <input aria-label="default time" [ngxTimepicker]="defaultValue" [value]="'05:11 pm'"  readonly>
             <ngx-material-timepicker #defaultValue></ngx-material-timepicker>
        </div>
-       
+
         <div class="ngx-material-timepicker-example__form-group">
             <input placeholder="Default time 11:11 pm" aria-label="default time"
                 [ngxTimepicker]="defaultTime" readonly>
             <ngx-material-timepicker #defaultTime [defaultTime]="'11:11 pm'"></ngx-material-timepicker>
         </div>
-       
+
        <div class="disabled-example">
             <input placeholder="Disabled Time Picker" aria-label="disabled time picker" [ngxTimepicker]="disabled" [disabled]="true">
             <ngx-material-timepicker #disabled></ngx-material-timepicker>
@@ -42,13 +42,13 @@ export class AppComponent {
                 [ngxTimepicker]="min" [min]="'03:15 am'" readonly>
             <ngx-material-timepicker #min></ngx-material-timepicker>
         </div>
-        
+
         <div class="max-time-example">
             <input placeholder="Max time 06:18 pm"
                 [ngxTimepicker]="max" max="06:18 pm" readonly>
             <ngx-material-timepicker #max></ngx-material-timepicker>
         </div>
-        
+
         <div class="time-range-example">
             <input placeholder="Time range from 12:10 am to 08:11 pm"
                 [ngxTimepicker]="range" min="12:10 am" max="08:11 pm" readonly>
@@ -61,7 +61,7 @@ export class AppComponent {
             <ngx-material-timepicker-toggle [for]="toggleTimepicker"></ngx-material-timepicker-toggle>
             <ngx-material-timepicker #toggleTimepicker></ngx-material-timepicker>
         </div>
-        
+
         <div class="toggle-custom-icon-example">
             <input [ngxTimepicker]="toggleIcon" [disableClick]="true" readonly>
             <ngx-material-timepicker-toggle [for]="toggleIcon">
@@ -71,7 +71,7 @@ export class AppComponent {
             </ngx-material-timepicker-toggle>
             <ngx-material-timepicker #toggleIcon></ngx-material-timepicker>
         </div>
-        
+
         <div class="toggle-disabled-example">
             <input [ngxTimepicker]="toggleTimepickerDisabled" [disableClick]="true" readonly>
             <ngx-material-timepicker-toggle [for]="toggleTimepickerDisabled" [disabled]="true"></ngx-material-timepicker-toggle>
@@ -93,12 +93,12 @@ export class AppComponent {
             [ngxTimepicker]="defaultGap" readonly>
         <ngx-material-timepicker #defaultGap></ngx-material-timepicker>
       </div>
-      
+
       <div class="minutes-gap-example">
         <input placeholder="Gap with 5"
             [ngxTimepicker]="with5Gap" readonly>
         <ngx-material-timepicker #with5Gap [minutesGap]="5"></ngx-material-timepicker>
-      </div>  
+      </div>
     `;
 
     customSettings = `
@@ -106,7 +106,7 @@ export class AppComponent {
             <input placeholder="Custom buttons" aria-label="Custom buttons" [ngxTimepicker]="timepickerWithButtons" readonly>
             <ngx-material-timepicker #timepickerWithButtons [cancelBtnTmpl]="cancelBtn"
                                              [confirmBtnTmpl]="confirmBtn"></ngx-material-timepicker>
-                                             
+
             <ng-template #cancelBtn>
                 <button style="margin-right: 10px;">Cancel tmpl</button>
             </ng-template>
@@ -114,13 +114,13 @@ export class AppComponent {
                 <button>Confirm tmpl</button>
             </ng-template>
          </div>
-         
+
          <div class="custom-theme-example">
             <input placeholder="Custom theme" aria-label="Custom theme"
                 [ngxTimepicker]="darkPicker" readonly>
             <ngx-material-timepicker #darkPicker [ngxMaterialTimepickerTheme]="darkTheme"></ngx-material-timepicker>
          </div>
-         
+
         //...
         darkTheme: NgxMaterialTimepickerTheme = {
             container: {
@@ -136,6 +136,16 @@ export class AppComponent {
                 clockFaceTimeInactiveColor: '#fff'
             }
         };
+
+        <div class="custom-hint-example">
+            <input placeholder="Custom hint" aria-label="Custom hint"
+                [ngxTimepicker]="customHint" readonly>
+            <ngx-material-timepicker #customHint [enableKeyboardInput]="true"
+                [editableHintTmpl]="dialHint"></ngx-material-timepicker>
+            <ng-template #dialHint>
+                <p>Custom hint</p>
+            </ng-template>
+        </div>
     `;
 
     connectedToInputSettings = `
@@ -160,4 +170,36 @@ export class AppComponent {
             clockFaceTimeInactiveColor: '#fff'
         }
     };
+
+    controlSimpleSettings = `
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [defaultTime]="'11:11 am'"></ngx-timepicker-field>
+        </div>
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [format]="24" [defaultTime]="'23:11'"></ngx-timepicker-field>
+        </div>
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [format]="24" [defaultTime]="'23:11'"></ngx-timepicker-field>
+        </div>
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [disabled]="true"></ngx-timepicker-field>
+        </div>
+    `;
+
+    toggleSettings = `
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [buttonAlign]="'left'"></ngx-timepicker-field>
+        </div>
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [toggleIcon]="icon"></ngx-timepicker-field>
+            <ng-template #icon>
+                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30" width="30px" height="30px">
+                    <path d="M15,3C8.373,3,3,8.373,3,15c0,6.627,5.373,12,12,12s12-5.373,12-12C27,8.373,21.627,3,15,3z M16,16H7.995 C7.445,16,7,15.555,7,15.005v-0.011C7,14.445,7.445,14,7.995,14H14V5.995C14,5.445,14.445,5,14.995,5h0.011 C15.555,5,16,5.445,16,5.995V16z"/>
+                </svg>
+            </ng-template>
+        </div>
+        <div class="ngx-timepicker-field-example">
+            <ngx-timepicker-field [controlOnly]="true"></ngx-timepicker-field>
+        </div>
+    `;
 }
