@@ -1,11 +1,11 @@
-import {BrowserModule} from '@angular/platform-browser';
-import {NgModule} from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
 
 
-import {AppComponent} from './app.component';
-import {ExampleSourceCodeComponent} from './example-source-code/example-source-code.component';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {NgxMaterialTimepickerModule} from './material-timepicker/ngx-material-timepicker.module';
+import { AppComponent } from './app.component';
+import { ExampleSourceCodeComponent } from './example-source-code/example-source-code.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgxMaterialTimepickerModule } from './material-timepicker/ngx-material-timepicker.module';
 
 
 @NgModule({
@@ -16,7 +16,7 @@ import {NgxMaterialTimepickerModule} from './material-timepicker/ngx-material-ti
     imports: [
         BrowserModule,
         BrowserAnimationsModule,
-        NgxMaterialTimepickerModule.forRoot()
+        NgxMaterialTimepickerModule,
     ],
     providers: [],
     bootstrap: [AppComponent]
