@@ -1,11 +1,10 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {
-    NgxMaterialTimepicker12HoursFaceComponent
-} from './ngx-material-timepicker-12-hours-face.component';
-import * as moment from 'moment';
-import {NO_ERRORS_SCHEMA, SimpleChanges} from '@angular/core';
-import {TimePeriod} from '../../models/time-period.enum';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NgxMaterialTimepicker12HoursFaceComponent } from './ngx-material-timepicker-12-hours-face.component';
+import { NO_ERRORS_SCHEMA, SimpleChanges } from '@angular/core';
+import { TimePeriod } from '../../models/time-period.enum';
+import { DateTime } from 'luxon';
+import * as TimepickerTime from '../../utils/timepicker-time.utils';
+import { spyOnFunction } from '../../ngx-material-timepicker.component.spec';
 
 describe('NgxMaterialTimepicker12HoursFaceComponent', () => {
     let fixture: ComponentFixture<NgxMaterialTimepicker12HoursFaceComponent>;
@@ -21,7 +20,7 @@ describe('NgxMaterialTimepicker12HoursFaceComponent', () => {
     });
 
     it('should call disabledHours once period changed', () => {
-        const spy = spyOn(TimepickerTime, 'disableHours');
+        const spy = spyOnFunction(TimepickerTime, 'disableHours');
         const changes: SimpleChanges = {
             period: {
                 currentValue: TimePeriod.PM,
@@ -30,7 +29,7 @@ describe('NgxMaterialTimepicker12HoursFaceComponent', () => {
                 isFirstChange: () => null
             }
         };
-        const time = moment();
+        const time = DateTime.fromJSDate(new Date());
         const format = 12;
         const period = TimePeriod.PM;
         const hours = TimepickerTime.getHours(format);
@@ -45,7 +44,7 @@ describe('NgxMaterialTimepicker12HoursFaceComponent', () => {
     });
 
     it('should not call disabledHours', () => {
-        const spy = spyOn(TimepickerTime, 'disableHours');
+        const spy = spyOnFunction(TimepickerTime, 'disableHours');
         const changes: SimpleChanges = {
             minTime: {
                 currentValue: null,
