@@ -1,7 +1,7 @@
 import {Component, Input, OnChanges, SimpleChanges} from '@angular/core';
 import {NgxMaterialTimepickerHoursFace} from '../timepicker-hours-face/ngx-material-timepicker-hours-face';
 import {TimePeriod} from '../../models/time-period.enum';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { disableHours } from '../../utils/timepicker-time.utils';
 
 
 @Component({
@@ -19,7 +19,7 @@ export class NgxMaterialTimepicker12HoursFaceComponent extends NgxMaterialTimepi
 
     ngOnChanges(changes: SimpleChanges) {
         if (changes['period'] && changes['period'].currentValue) {
-            this.hoursList = TimepickerTime.disableHours(this.hoursList, {
+            this.hoursList = disableHours(this.hoursList, {
                 min: this.minTime,
                 max: this.maxTime,
                 format: this.format,
