@@ -1,8 +1,9 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {NO_ERRORS_SCHEMA} from '@angular/core';
-import {NgxMaterialTimepicker24HoursFaceComponent} from './ngx-material-timepicker-24-hours-face.component';
-import {TimepickerTime} from '../../timepicker-time.namespace';
-import * as moment from 'moment';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { NgxMaterialTimepicker24HoursFaceComponent } from './ngx-material-timepicker-24-hours-face.component';
+import * as TimepickerTime from '../../utils/timepicker-time.utils';
+import { DateTime } from 'luxon';
+import { spyOnFunction } from '../../ngx-material-timepicker.component.spec';
 
 describe('NgxMaterialTimepicker24HoursFaceComponent', () => {
     let fixture: ComponentFixture<NgxMaterialTimepicker24HoursFaceComponent>;
@@ -18,8 +19,8 @@ describe('NgxMaterialTimepicker24HoursFaceComponent', () => {
     });
 
     it('should call disableHours', () => {
-        const spy = spyOn(TimepickerTime, 'disableHours');
-        const time = moment();
+        const spy = spyOnFunction(TimepickerTime, 'disableHours');
+        const time = DateTime.fromJSDate(new Date());
         const format = 24;
         const hours = TimepickerTime.getHours(format);
 
