@@ -1,6 +1,6 @@
 import {AfterContentInit, Component} from '@angular/core';
 import {NgxMaterialTimepickerHoursFace} from '../timepicker-hours-face/ngx-material-timepicker-hours-face';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { disableHours } from '../../utils/timepicker-time.utils';
 
 @Component({
     selector: 'ngx-material-timepicker-24-hours-face',
@@ -14,7 +14,7 @@ export class NgxMaterialTimepicker24HoursFaceComponent extends NgxMaterialTimepi
     }
 
     ngAfterContentInit() {
-        this.hoursList = TimepickerTime.disableHours(this.hoursList, {
+        this.hoursList = disableHours(this.hoursList, {
             min: this.minTime,
             max: this.maxTime,
             format: this.format
