@@ -1,8 +1,8 @@
-import {ComponentFixture, fakeAsync, TestBed, tick} from '@angular/core/testing';
-import {NgxMaterialTimepickerDialControlComponent} from './ngx-material-timepicker-dial-control.component';
-import {NO_ERRORS_SCHEMA, SimpleChanges} from '@angular/core';
-import {TimeUnit} from '../../models/time-unit.enum';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { NgxMaterialTimepickerDialControlComponent } from './ngx-material-timepicker-dial-control.component';
+import { NO_ERRORS_SCHEMA, SimpleChanges } from '@angular/core';
+import { TimeUnit } from '../../models/time-unit.enum';
+import { getHours } from '../../utils/timepicker-time.utils';
 
 describe('NgxMaterialTimepickerDialControlComponent', () => {
     let fixture: ComponentFixture<NgxMaterialTimepickerDialControlComponent>;
@@ -17,9 +17,10 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
         component = fixture.componentInstance;
     });
 
-    it('should set current time to previous time and change time unit', fakeAsync(() => {
-        let timeUnit = null;
-        component.timeUnitChanged.subscribe(unit => timeUnit = unit);
+    it('should set current time to previous time, change time unit and emit focus event', async (() => {
+        let counter = 0;
+        component.timeUnitChanged.subscribe(unit => expect(unit).toBe(TimeUnit.MINUTE));
+        component.focused.subscribe(() => expect(++counter).toBe(1));
 
         component.time = '10';
         expect(component.previousTime).toBeUndefined();
@@ -27,7 +28,6 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
         component.saveTimeAndChangeTimeUnit({preventDefault: () => null} as FocusEvent, TimeUnit.MINUTE);
 
         expect(component.previousTime).toBe('10');
-        expect(timeUnit).toBe(TimeUnit.MINUTE);
     }));
 
     it('should emit changed time if it exists and available', fakeAsync(() => {
@@ -102,7 +102,9 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
         expect(component.time).toBe('4');
     });
 
-    it('should format time if editable', () => {
+    it('should format time if editable and emit unfocused event', async(() => {
+        let counter = 0;
+
         component.isEditable = true;
         component.time = '2';
         component.previousTime = 4;
@@ -111,83 +113,28 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
         component.formatTime();
         expect(component.time).toBe('02');
 
+        component.unfocused.subscribe(() => expect(++counter).toBe(1));
         component.time = '';
         component.formatTime();
-        expect(component.time).toBe('04');
 
+        expect(component.time).toBe('04');
         component.time = '5';
         component.isEditable = false;
         component.formatTime();
         expect(component.time).toBe('5');
-    });
+    }));
 
     describe('onKeyDown', () => {
         let counter = 0;
         const event = {
             keyCode: 0, preventDefault: () => {
-                counter++
+                counter++;
             }
         } as KeyboardEvent;
-        const numbers = Array(10).fill(48).map((v, i) => v + i);
-        const numpadNumbers = Array(10).fill(96).map((v, i) => v + i);
-        const arrows = Array(6).fill(35).map((v, i) => v + i); // home, end, left, right, up, down
-        const specialKeys = [46, 8, 9, 27, 13]; // backspace, delete, tab, escape, enter
 
         beforeEach(() => {
             counter = 0;
-            component.timeList = TimepickerTime.getHours(24);
-        });
-
-
-        it('should allow numbers', () => {
-
-            const keyCodes = numbers.concat(numpadNumbers);
-            component.time = '';
-
-
-            keyCodes.forEach(code => {
-                component.onKeyDown({...event, keyCode: code});
-                expect(counter).toBe(0);
-            });
-        });
-
-        it('should allow backspace, delete, tab, escape, enter', () => {
-            specialKeys.forEach(code => {
-                component.onKeyDown({...event, keyCode: code});
-                expect(counter).toBe(0);
-            });
-        });
-
-        it('should allow home, end, left, right, up, down', () => {
-            arrows.forEach(code => {
-                component.onKeyDown({...event, keyCode: code});
-                expect(counter).toBe(0);
-            });
-        });
-
-        it('should allow ctrl/cmd+a, ctrl/cmd+c, ctrl/cmd+x', () => {
-            const chars = [65, 67, 88];
-
-            chars.forEach(code => {
-                component.onKeyDown({...event, keyCode: code, ctrlKey: true});
-                expect(counter).toBe(0);
-            });
-
-            chars.forEach(code => {
-                component.onKeyDown({...event, keyCode: code, metaKey: true});
-                expect(counter).toBe(0);
-            });
-        });
-
-        it('should not allow chars but numbers, backspace, delete, tab, escape, enter, home, end, left, right, up, down', () => {
-            const allKeyCodes = Array(114).fill(8).map((v, i) => v + i);
-            const allowedCodes = [...numbers, ...numpadNumbers, ...specialKeys, ...arrows];
-            const restrictedCodes = allKeyCodes.filter(code => !allowedCodes.includes(code));
-
-            restrictedCodes.forEach((code, index) => {
-                component.onKeyDown({...event, keyCode: code});
-                expect(counter).toBe(index + 1);
-            })
+            component.timeList = getHours(24);
         });
 
         it('should call preventDefault if no time exist or time disabled', () => {
@@ -201,7 +148,7 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
 
             component.time = '';
             component.onKeyDown({...event, keyCode: NUM_1});
-            expect(counter).toBe(2)
+            expect(counter).toBe(2);
         });
 
         it('should up time by 1', () => {
@@ -218,6 +165,24 @@ describe('NgxMaterialTimepickerDialControlComponent', () => {
 
             component.onKeyDown({...event, keyCode: ARROW_DOWN});
             expect(component.time).toBe('10');
+        });
+
+        it('should up time by 7', () => {
+            const ARROW_UP = 38;
+            component.time = '11';
+            component.minutesGap = 7;
+
+            component.onKeyDown({...event, keyCode: ARROW_UP});
+            expect(component.time).toBe('18');
+        });
+
+        it('should down time by 6', () => {
+            const ARROW_DOWN = 40;
+            component.time = '11';
+            component.minutesGap = 6;
+
+            component.onKeyDown({...event, keyCode: ARROW_DOWN});
+            expect(component.time).toBe('5');
         });
     });
 });
