@@ -1,7 +1,9 @@
-import {Component, EventEmitter, Input, OnChanges, Output, SimpleChanges} from '@angular/core';
-import {ClockFaceTime} from '../../models/clock-face-time.interface';
-import {TimeUnit} from '../../models/time-unit.enum';
-import {TimeFormatterPipe} from '../../pipes/time-formatter.pipe';
+/* tslint:disable:triple-equals */
+import { Component, EventEmitter, Input, OnChanges, Output, SimpleChanges } from '@angular/core';
+import { ClockFaceTime } from '../../models/clock-face-time.interface';
+import { TimeUnit } from '../../models/time-unit.enum';
+import { TimeFormatterPipe } from '../../pipes/time-formatter.pipe';
+import { isDigit } from '../../utils/timepicker.utils';
 
 @Component({
     selector: 'ngx-material-timepicker-dial-control',
@@ -17,9 +19,12 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
     @Input() time: string;
     @Input() isActive: boolean;
     @Input() isEditable: boolean;
+    @Input() minutesGap: number;
 
     @Output() timeUnitChanged = new EventEmitter<TimeUnit>();
     @Output() timeChanged = new EventEmitter<ClockFaceTime>();
+    @Output() focused = new EventEmitter<null>();
+    @Output() unfocused = new EventEmitter<null>();
 
     private get selectedTime(): ClockFaceTime {
         if (!!this.time) {
@@ -32,7 +37,7 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
             if (this.isEditable && !changes['time'].firstChange) {
                 return;
             }
-            this.time = new TimeFormatterPipe().transform(+changes['time'].currentValue, this.timeUnit)
+            this.time = new TimeFormatterPipe().transform(+changes['time'].currentValue, this.timeUnit);
         }
     }
 
@@ -40,6 +45,7 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
         event.preventDefault();
         this.previousTime = this.time;
         this.timeUnitChanged.next(unit);
+        this.focused.next();
     }
 
     updateTime(): void {
@@ -54,6 +60,7 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
         if (this.isEditable) {
             const time = this.time || this.previousTime;
             this.time = new TimeFormatterPipe().transform(+time, this.timeUnit);
+            this.unfocused.next();
         }
     }
 
@@ -61,11 +68,11 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
         const char = String.fromCharCode(e.keyCode);
 
 
-        if ((!isInputAllowed(e)) || isTimeDisabledToChange(this.time, char, this.timeList)) {
+        if ((!isDigit(e)) || isTimeDisabledToChange(this.time, char, this.timeList)) {
             e.preventDefault();
         }
 
-        if (isInputAllowed(e)) {
+        if (isDigit(e)) {
             this.changeTimeByArrow(e.keyCode);
         }
     }
@@ -76,9 +83,9 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
         let time: string;
 
         if (keyCode === ARROW_UP) {
-            time = String(+this.time + 1);
+            time = String(+this.time + (this.minutesGap || 1));
         } else if (keyCode === ARROW_DOWN) {
-            time = String(+this.time - 1);
+            time = String(+this.time - (this.minutesGap || 1));
         }
 
         if (!isTimeUnavailable(time, this.timeList)) {
@@ -89,22 +96,6 @@ export class NgxMaterialTimepickerDialControlComponent implements OnChanges {
 
 }
 
-function isInputAllowed(e: KeyboardEvent): boolean {
-    // Allow: backspace, delete, tab, escape, enter
-    if ([46, 8, 9, 27, 13].some(n => n === e.keyCode) ||
-        // Allow: Ctrl/cmd+A
-        (e.keyCode == 65 && (e.ctrlKey === true || e.metaKey === true)) ||
-        // Allow: Ctrl/cmd+C
-        (e.keyCode == 67 && (e.ctrlKey === true || e.metaKey === true)) ||
-        // Allow: Ctrl/cmd+X
-        (e.keyCode == 88 && (e.ctrlKey === true || e.metaKey === true)) ||
-        // Allow: home, end, left, right, up, down
-        (e.keyCode >= 35 && e.keyCode <= 40)) {
-
-        return true;
-    }
-    return !((e.keyCode < 48 || e.keyCode > 57) && (e.keyCode < 96 || e.keyCode > 105))
-}
 
 function isTimeDisabledToChange(currentTime: string, nextTime: string, timeList: ClockFaceTime[]): boolean {
     const isNumber = /\d/.test(nextTime);
