@@ -1,4 +1,4 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import {Component} from '@angular/core';
 import {NgxMaterialTimepickerHoursFace} from './ngx-material-timepicker-hours-face';
@@ -42,4 +42,18 @@ describe('NgxMaterialTimepickerHoursFace', () => {
     it('should generate array with 24 items', () => {
         expect(component24.hoursList.length).toBe(24);
     });
+
+    it('should emit selected hour (12hr format)', async(() => {
+        const time = 10;
+
+        component12.hourSelected.subscribe(hour => expect(hour).toBe(time));
+        component12.onTimeSelected(time);
+    }));
+
+    it('should emit selected hour (24hr format)', async(() => {
+        const time = 15;
+
+        component24.hourSelected.subscribe(hour => expect(hour).toBe(time));
+        component24.onTimeSelected(time);
+    }));
 });
