@@ -1,22 +1,26 @@
-import {EventEmitter, Input, Output} from '@angular/core';
-import {ClockFaceTime} from '../../models/clock-face-time.interface';
-import {Moment} from 'moment';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { EventEmitter, Input, Output } from '@angular/core';
+import { DateTime } from 'luxon';
+import { ClockFaceTime } from '../../models/clock-face-time.interface';
+import { getHours } from '../../utils/timepicker-time.utils';
 
 
 export class NgxMaterialTimepickerHoursFace {
 
     @Input() selectedHour: ClockFaceTime;
-    @Input() minTime: Moment;
-    @Input() maxTime: Moment;
+    @Input() minTime: DateTime;
+    @Input() maxTime: DateTime;
     @Input() format: number;
 
     @Output() hourChange = new EventEmitter<ClockFaceTime>();
-    @Output() hourSelected = new EventEmitter<null>();
+    @Output() hourSelected = new EventEmitter<number>();
 
     hoursList: ClockFaceTime[] = [];
 
     protected constructor(format: number) {
-        this.hoursList = TimepickerTime.getHours(format);
+        this.hoursList = getHours(format);
+    }
+
+    onTimeSelected(time: number): void {
+        this.hourSelected.next(time);
     }
 }
