@@ -1,9 +1,10 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {NgxMaterialTimepickerMinutesFaceComponent} from './ngx-material-timepicker-minutes-face.component';
-import {NO_ERRORS_SCHEMA, SimpleChanges} from '@angular/core';
-import * as moment from 'moment';
-import {TimePeriod} from '../../models/time-period.enum';
-import {TimepickerTime} from '../../timepicker-time.namespace';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NgxMaterialTimepickerMinutesFaceComponent } from './ngx-material-timepicker-minutes-face.component';
+import { NO_ERRORS_SCHEMA, SimpleChanges } from '@angular/core';
+import { TimePeriod } from '../../models/time-period.enum';
+import * as TimepickerTime from '../../utils/timepicker-time.utils';
+import { DateTime } from 'luxon';
+import { spyOnFunction } from '../../ngx-material-timepicker.component.spec';
 
 describe('NgxMaterialTimepickerMinutesFaceComponent', () => {
     let fixture: ComponentFixture<NgxMaterialTimepickerMinutesFaceComponent>;
@@ -19,7 +20,7 @@ describe('NgxMaterialTimepickerMinutesFaceComponent', () => {
     });
 
     it('should call disableMinutes once period changed', () => {
-        const spy = spyOn(TimepickerTime, 'disableMinutes');
+        const spy = spyOnFunction(TimepickerTime, 'disableMinutes');
         const changes: SimpleChanges = {
             period: {
                 currentValue: TimePeriod.PM,
@@ -28,7 +29,7 @@ describe('NgxMaterialTimepickerMinutesFaceComponent', () => {
                 isFirstChange: () => null
             }
         };
-        const time = moment();
+        const time = DateTime.fromJSDate(new Date());
         const format = 12;
         const period = TimePeriod.PM;
         const minutes = TimepickerTime.getMinutes();
@@ -44,7 +45,7 @@ describe('NgxMaterialTimepickerMinutesFaceComponent', () => {
     });
 
     it('should not call disableMinutes', () => {
-        const spy = spyOn(TimepickerTime, 'disableMinutes');
+        const spy = spyOnFunction(TimepickerTime, 'disableMinutes');
         const changes: SimpleChanges = {
             minTime: {
                 currentValue: null,
