@@ -1,10 +1,10 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {TimePeriod} from '../../models/time-period.enum';
-import {TimeUnit} from '../../models/time-unit.enum';
-import {ClockFaceTime} from '../../models/clock-face-time.interface';
-import {TimepickerTime} from '../../timepicker-time.namespace';
-import {Moment} from 'moment';
-import {animate, sequence, style, transition, trigger} from '@angular/animations';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { TimePeriod } from '../../models/time-period.enum';
+import { TimeUnit } from '../../models/time-unit.enum';
+import { ClockFaceTime } from '../../models/clock-face-time.interface';
+import { animate, sequence, style, transition, trigger } from '@angular/animations';
+import { DateTime } from 'luxon';
+import { disableHours, disableMinutes } from '../../utils/timepicker-time.utils';
 
 @Component({
     selector: 'ngx-material-timepicker-period',
@@ -33,8 +33,8 @@ export class NgxMaterialTimepickerPeriodComponent {
     @Input() activeTimeUnit: TimeUnit;
     @Input() hours: ClockFaceTime[];
     @Input() minutes: ClockFaceTime[];
-    @Input() minTime: Moment;
-    @Input() maxTime: Moment;
+    @Input() minTime: DateTime;
+    @Input() maxTime: DateTime;
     @Input() selectedHour: number | string;
 
     @Output() periodChanged = new EventEmitter<TimePeriod>();
@@ -58,14 +58,14 @@ export class NgxMaterialTimepickerPeriodComponent {
     private getDisabledTimeByPeriod(period: TimePeriod): ClockFaceTime[] {
         switch (this.activeTimeUnit) {
             case TimeUnit.HOUR:
-                return TimepickerTime.disableHours(this.hours, {
+                return disableHours(this.hours, {
                     min: this.minTime,
                     max: this.maxTime,
                     format: this.format,
                     period
                 });
             case TimeUnit.MINUTE:
-                return TimepickerTime.disableMinutes(this.minutes, +this.selectedHour, {
+                return disableMinutes(this.minutes, +this.selectedHour, {
                     min: this.minTime,
                     max: this.maxTime,
                     format: this.format,
