@@ -1,13 +1,10 @@
-import {Component, DebugElement, SimpleChanges} from '@angular/core';
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {TimepickerDirective} from './ngx-timepicker.directive';
-import {By} from '@angular/platform-browser';
-import {NgxMaterialTimepickerComponent} from '../ngx-material-timepicker.component';
-import {NgxMaterialTimepickerModule} from '../ngx-material-timepicker.module';
-import * as _moment from 'moment';
-
-const moment = _moment;
-
+import { Component, DebugElement, SimpleChanges } from '@angular/core';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { TimepickerDirective } from './ngx-timepicker.directive';
+import { By } from '@angular/platform-browser';
+import { NgxMaterialTimepickerComponent } from '../ngx-material-timepicker.component';
+import { NgxMaterialTimepickerModule } from '../ngx-material-timepicker.module';
+import { DateTime } from 'luxon';
 
 @Component({
     template: `
@@ -31,7 +28,7 @@ describe('TimepickerDirective', () => {
             declarations: [
                 TestComponent
             ],
-            imports: [NgxMaterialTimepickerModule.forRoot()]
+            imports: [NgxMaterialTimepickerModule]
         }).createComponent(TestComponent);
         input = fixture.debugElement.query(By.directive(TimepickerDirective));
         directive = input.injector.get<TimepickerDirective>(TimepickerDirective);
@@ -60,28 +57,28 @@ describe('TimepickerDirective', () => {
         expect(directive.format).toBe(24);
     });
 
-    it('should return min time in Moment type if pass string', () => {
+    it('should return min time in DateTime type if pass string', () => {
         directive.min = '11:00 pm';
-        expect(directive.min['hour']()).toBe(23);
-        expect(directive.min['minute']()).toBe(0);
+        expect(directive.min['hour']).toBe(23);
+        expect(directive.min['minute']).toBe(0);
     });
 
-    it('should return min time in Moment type if pass moment', () => {
-        directive.min = moment().hour(10).minute(11);
-        expect(directive.min['hour']()).toBe(10);
-        expect(directive.min['minute']()).toBe(11);
+    it('should return min time in DateTime type if pass DateTime', () => {
+        directive.min = DateTime.fromObject({hour: 10, minute: 11});
+        expect(directive.min.hour).toBe(10);
+        expect(directive.min.minute).toBe(11);
     });
 
-    it('should return max time in Moment type if pass string', () => {
+    it('should return max time in DateTime type if pass string', () => {
         directive.max = '11:00 pm';
-        expect(directive.max['hour']()).toBe(23);
-        expect(directive.max['minute']()).toBe(0);
+        expect(directive.max['hour']).toBe(23);
+        expect(directive.max['minute']).toBe(0);
     });
 
-    it('should return max time in Moment type if pass moment', () => {
-        directive.max = moment().hour(10).minute(11);
-        expect(directive.max['hour']()).toBe(10);
-        expect(directive.max['minute']()).toBe(11);
+    it('should return max time in DateTime type if pass DateTime', () => {
+        directive.max = DateTime.fromObject({hour: 10, minute: 11});
+        expect(directive.max.hour).toBe(10);
+        expect(directive.max.minute).toBe(11);
     });
 
     it(`should clear the time if set value undefined, null, '' `, () => {
@@ -94,11 +91,13 @@ describe('TimepickerDirective', () => {
     });
 
     it('should return formatted time', () => {
+        directive.timepicker = timepickerComponent;
         directive.value = '11:00';
         expect(directive.value).toBe('11:00 am');
     });
 
     it('should call console.warn if time is not between min and max(inclusively) value', () => {
+        directive.timepicker = timepickerComponent;
         const spy = spyOn(console, 'warn');
         directive.min = '11:00 am';
         directive.value = '10:00 am';
@@ -127,13 +126,15 @@ describe('TimepickerDirective', () => {
     });
 
     it('should change time onInput', () => {
+        directive.timepicker = timepickerComponent;
         directive.onInput('11:12');
         expect(directive.value).toBe('11:12 am');
     });
 
-    it('should set Invalid date if time is in inappropriate format', () => {
+    it('should set invalid datetime if time is in inappropriate format', () => {
+        directive.timepicker = timepickerComponent;
         directive.value = 'test';
-        expect(directive.value).toBe('Invalid date');
+        expect(directive.value).toBe('invalid datetime');
     });
 
     it('should set default time if binding value changes', () => {
@@ -196,6 +197,7 @@ describe('TimepickerDirective', () => {
     });
 
     it('should set onChange function on registerOnChange', () => {
+        directive.timepicker = timepickerComponent;
         const spy = spyOn(console, 'log');
         const time = '11:12 am';
 
