@@ -1,12 +1,13 @@
-import {Directive, ElementRef, forwardRef, HostListener, Input, OnChanges, OnDestroy, SimpleChanges} from '@angular/core';
-import {NgxMaterialTimepickerComponent} from '../ngx-material-timepicker.component';
-import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
-import {Subscription} from 'rxjs';
-import {Moment} from 'moment';
-import {TimeAdapter} from '../services/time-adapter';
+import { Directive, ElementRef, forwardRef, HostListener, Input, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
+import { NgxMaterialTimepickerComponent } from '../ngx-material-timepicker.component';
+import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { Subscription } from 'rxjs';
+import { TimeAdapter } from '../services/time-adapter';
+import { DateTime } from 'luxon';
 
 const VALUE_ACCESSOR = {
     provide: NG_VALUE_ACCESSOR,
+    // tslint:disable-next-line
     useExisting: forwardRef(() => TimepickerDirective),
     multi: true
 };
@@ -34,34 +35,41 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
     private _format = 12;
 
     @Input()
-    set min(value: string | Moment) {
+    set min(value: string | DateTime) {
         if (typeof value === 'string') {
-            this._min = TimeAdapter.convertTimeToMoment(value);
+            this._min = TimeAdapter.convertTimeToDateTime(value, this._format);
             return;
         }
         this._min = value;
     }
 
-    get min(): string | Moment {
+    get min(): string | DateTime {
         return this._min;
     }
 
-    private _min: string | Moment;
+    private _min: string | DateTime;
 
     @Input()
-    set max(value: string | Moment) {
+    set max(value: string | DateTime) {
         if (typeof value === 'string') {
-            this._max = TimeAdapter.convertTimeToMoment(value);
+            this._max = TimeAdapter.convertTimeToDateTime(value, this._format);
             return;
         }
         this._max = value;
     }
 
-    get max(): string | Moment {
+    get max(): string | DateTime {
         return this._max;
     }
 
-    private _max: string | Moment;
+    private _max: string | DateTime;
+
+    @Input('ngxTimepicker')
+    set timepicker(picker: NgxMaterialTimepickerComponent) {
+        this.registerTimepicker(picker);
+    }
+
+    private _timepicker: NgxMaterialTimepickerComponent;
 
     @Input()
     set value(value: string) {
@@ -71,7 +79,16 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
             return;
         }
         const time = TimeAdapter.formatTime(value, this._format);
-        if (TimeAdapter.isTimeAvailable(time, <Moment>this._min, <Moment>this._max, 'minutes')) {
+        const isAvailable = TimeAdapter.isTimeAvailable(
+            time,
+            <DateTime>this._min,
+            <DateTime>this._max,
+            'minutes',
+            this._timepicker.minutesGap,
+            this._format
+        );
+
+        if (isAvailable) {
             this._value = time;
             this.updateInputValue();
             return;
@@ -85,17 +102,16 @@ export class TimepickerDirective implements ControlValueAccessor, OnDestroy, OnC
 
     private _value = '';
 
-    @Input('ngxTimepicker')
-    set timepicker(picker: NgxMaterialTimepickerComponent) {
-        this.registerTimepicker(picker);
-    }
-    private _timepicker: NgxMaterialTimepickerComponent;
-
     @Input() disabled: boolean;
     @Input() disableClick: boolean;
+
     private timepickerSubscriptions: Subscription[] = [];
-    onTouched = () => {};
-    private onChange: (value: any) => void = () => {};
+
+    onTouched = () => {
+    }
+
+    private onChange: (value: any) => void = () => {
+    }
 
     constructor(private elementRef: ElementRef) {
     }
