@@ -1,4 +1,6 @@
 export enum TimeFormat {
     TWELVE = 'hh:mm a',
-    TWENTY_FOUR = 'HH:mm'
+    TWELVE_SHORT = 'h:m a',
+    TWENTY_FOUR = 'HH:mm',
+    TWENTY_FOUR_SHORT = 'H:m'
 }
