@@ -1,21 +1,21 @@
-import {Pipe, PipeTransform} from '@angular/core';
-import * as moment from 'moment';
-import {TimeUnit} from '../models/time-unit.enum';
+import { Pipe, PipeTransform } from '@angular/core';
+import { TimeUnit } from '../models/time-unit.enum';
+import { DateTime } from 'luxon';
 
 @Pipe({
     name: 'timeFormatter'
 })
 export class TimeFormatterPipe implements PipeTransform {
 
-    transform(time: number, timeUnit: TimeUnit): any {
-        if (time === undefined) {
+    transform(time: number | string, timeUnit: TimeUnit): any {
+        if (time == null || time === '') {
             return time;
         }
         switch (timeUnit) {
             case TimeUnit.HOUR:
-                return moment.utc(time * 3600 * 1000).format('HH');
+                return DateTime.fromObject({hour: +time}).toFormat('HH');
             case TimeUnit.MINUTE:
-                return moment.utc(time * 60 * 1000).format('mm');
+                return DateTime.fromObject({minute: +time}).toFormat('mm');
             default:
                 throw new Error('no such time unit');
         }
