@@ -1,7 +1,9 @@
 import {Injectable} from '@angular/core';
 import {Observable, Subject} from 'rxjs';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class NgxMaterialTimepickerEventService {
 
     private backdropClickSubject: Subject<MouseEvent> = new Subject();
