@@ -1,8 +1,9 @@
-import {TestBed} from '@angular/core/testing';
-import {ClockFaceTime} from '../models/clock-face-time.interface';
-import {NgxMaterialTimepickerService} from './ngx-material-timepicker.service';
-import {TimePeriod} from '../models/time-period.enum';
-import * as moment from 'moment';
+import { TestBed } from '@angular/core/testing';
+import { ClockFaceTime } from '../models/clock-face-time.interface';
+import { NgxMaterialTimepickerService } from './ngx-material-timepicker.service';
+import { TimePeriod } from '../models/time-period.enum';
+import { TimeAdapter } from './time-adapter';
+import { DateTime } from 'luxon';
 
 describe('NgxMaterialTimepickerService', () => {
     const DEFAULT_HOUR: ClockFaceTime = {
@@ -55,7 +56,7 @@ describe('NgxMaterialTimepickerService', () => {
         expect(selectedPeriod).toEqual(TimePeriod.PM);
     });
 
-    it('should return full time as string (hh:mm a or HH:mm)', () => {
+    it('should return default full time as string (hh:mm a or HH:mm)', () => {
         expect(timepickerService.getFullTime(12)).toBe('12:00 am');
         expect(timepickerService.getFullTime(24)).toBe('12:00');
     });
@@ -68,11 +69,37 @@ describe('NgxMaterialTimepickerService', () => {
         expect(selectedMinute).toEqual({...DEFAULT_MINUTE, time: 15});
         expect(selectedPeriod).toBe(TimePeriod.AM);
 
+        time = '11:12 pm';
+        timepickerService.setDefaultTimeIfAvailable(time, null, null, 12);
+        expect(selectedHour.time).toBe(11);
+        expect(selectedMinute.time).toBe(12);
+        expect(selectedPeriod).toBe(TimePeriod.PM);
+
+        time = '12:00 pm';
+        timepickerService.setDefaultTimeIfAvailable(time, null, null, 12);
+        expect(selectedHour.time).toBe(12);
+        expect(selectedMinute.time).toBe(0);
+        expect(selectedPeriod).toBe(TimePeriod.PM);
+
+        time = '12:00 am';
+        timepickerService.setDefaultTimeIfAvailable(time, null, null, 12);
+        expect(selectedHour.time).toBe(12);
+        expect(selectedMinute.time).toBe(0);
+        expect(selectedPeriod).toBe(TimePeriod.AM);
+
         time = '00:00';
         timepickerService.setDefaultTimeIfAvailable(time, null, null, 24);
 
         expect(selectedHour).toEqual({...DEFAULT_HOUR, time: 0});
         expect(selectedMinute).toEqual({...DEFAULT_MINUTE, time: 0});
+        expect(selectedPeriod).toBe(TimePeriod.AM);
+
+        time = '15:00';
+        timepickerService.setDefaultTimeIfAvailable(time, null, null, 24);
+
+        expect(selectedHour).toEqual({...DEFAULT_HOUR, time: 15});
+        expect(selectedMinute).toEqual({...DEFAULT_MINUTE, time: 0});
+        expect(selectedPeriod).toBe(TimePeriod.AM);
     });
 
     it('should reset time if default time is invalid', () => {
@@ -90,7 +117,7 @@ describe('NgxMaterialTimepickerService', () => {
     });
 
     it('should not change time if it is not available', () => {
-        const min = moment().hour(11);
+        const min = DateTime.fromObject({hour: 11});
 
         timepickerService.setDefaultTimeIfAvailable('10:10 am', null, null, 12);
 
@@ -103,5 +130,15 @@ describe('NgxMaterialTimepickerService', () => {
         expect(selectedHour).toEqual({...DEFAULT_HOUR, time: 10});
         expect(selectedMinute).toEqual({...DEFAULT_MINUTE, time: 10});
         expect(selectedPeriod).toBe(TimePeriod.AM);
+    });
+
+    it('should call console error', () => {
+        const minutesGap = 5;
+        const min = TimeAdapter.convertTimeToDateTime('11:00 pm');
+        const max = TimeAdapter.convertTimeToDateTime('11:50 pm');
+        const spy = spyOn(console, 'error');
+
+        timepickerService.setDefaultTimeIfAvailable('11:43 pm', min, max, 12, minutesGap);
+        expect(spy).toHaveBeenCalled();
     });
 });
