@@ -1,13 +1,10 @@
-import {Injectable} from '@angular/core';
-import {ClockFaceTime} from '../models/clock-face-time.interface';
-import {BehaviorSubject, Observable} from 'rxjs';
-import {TimePeriod} from '../models/time-period.enum';
-import * as moment_ from 'moment';
-import {TimeFormat} from '../models/time-format.enum';
-import {TimeAdapter} from './time-adapter';
-import {Moment} from 'moment';
+import { Injectable } from '@angular/core';
+import { ClockFaceTime } from '../models/clock-face-time.interface';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { TimePeriod } from '../models/time-period.enum';
+import { TimeAdapter } from './time-adapter';
+import { DateTime } from 'luxon';
 
-const moment = moment_;
 
 const DEFAULT_HOUR: ClockFaceTime = {
     time: 12,
@@ -18,24 +15,15 @@ const DEFAULT_MINUTE: ClockFaceTime = {
     angle: 360
 };
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class NgxMaterialTimepickerService {
 
     private hourSubject = new BehaviorSubject<ClockFaceTime>(DEFAULT_HOUR);
     private minuteSubject = new BehaviorSubject<ClockFaceTime>(DEFAULT_MINUTE);
     private periodSubject = new BehaviorSubject<TimePeriod>(TimePeriod.AM);
 
-    private set defaultTime(time: string) {
-        const defaultTime = moment(time, TimeFormat.TWENTY_FOUR).toDate();
-
-        if (moment(defaultTime).isValid()) {
-            this.hour = {...DEFAULT_HOUR, time: defaultTime.getHours()};
-            this.minute = {...DEFAULT_MINUTE, time: defaultTime.getMinutes()};
-            this.period = <TimePeriod>time.substr(time.length - 2).toUpperCase();
-        } else {
-            this.resetTime();
-        }
-    }
 
     set hour(hour: ClockFaceTime) {
         this.hourSubject.next(hour);
@@ -54,7 +42,11 @@ export class NgxMaterialTimepickerService {
     }
 
     set period(period: TimePeriod) {
-        this.periodSubject.next(period);
+        const isPeriodValid = (period === TimePeriod.AM) || (period === TimePeriod.PM);
+
+        if (isPeriodValid) {
+            this.periodSubject.next(period);
+        }
     }
 
     get selectedPeriod(): Observable<TimePeriod> {
@@ -62,9 +54,14 @@ export class NgxMaterialTimepickerService {
     }
 
 
-    setDefaultTimeIfAvailable(time: string, min: Moment, max: Moment, format: number) {
-        if (TimeAdapter.isTimeAvailable(time, min, max, 'minutes')) {
-            this.defaultTime = TimeAdapter.formatTime(time, format);
+    setDefaultTimeIfAvailable(time: string, min: DateTime, max: DateTime, format: number, minutesGap?: number) {
+        /* Workaround to double error message*/
+        try {
+            if (TimeAdapter.isTimeAvailable(time, min, max, 'minutes', minutesGap)) {
+                this.setDefaultTime(TimeAdapter.formatTime(time, format), format);
+            }
+        } catch (e) {
+            console.error(e);
         }
     }
 
@@ -76,9 +73,39 @@ export class NgxMaterialTimepickerService {
         return TimeAdapter.formatTime(`${hour}:${minute} ${period}`, format);
     }
 
+    private setDefaultTime(time: string, format: number) {
+        const defaultTime = TimeAdapter.convertTimeToDateTime(time, format).toJSDate();
+
+        if (DateTime.fromJSDate(defaultTime).isValid) {
+            const period = time.substr(time.length - 2).toUpperCase();
+            const hour = defaultTime.getHours();
+
+            this.hour = {...DEFAULT_HOUR, time: formatHourByPeriod(hour, period as TimePeriod)};
+            this.minute = {...DEFAULT_MINUTE, time: defaultTime.getMinutes()};
+            this.period = period as TimePeriod;
+
+        } else {
+            this.resetTime();
+        }
+    }
+
     private resetTime(): void {
         this.hour = {...DEFAULT_HOUR};
         this.minute = {...DEFAULT_MINUTE};
         this.period = TimePeriod.AM;
+    }
+}
+
+/***
+ *  Format hour in 24hours format to meridian (AM or PM) format
+ */
+function formatHourByPeriod(hour: number, period: TimePeriod): number {
+    switch (period) {
+        case TimePeriod.AM:
+            return hour === 0 ? 12 : hour;
+        case TimePeriod.PM:
+            return hour === 12 ? 12 : hour - 12;
+        default:
+            return hour;
     }
 }
